+++ tests/uniswap_v2/test_uniswap_v2_analysis.py
@@ -119,7 +119,7 @@ def test_analyse_buy_success(web3: Web3, deployer: str, user_1, uniswap_v2: Unis
 
     analysis = analyse_trade_by_hash(web3, uniswap_v2, tx_hash)
     assert isinstance(analysis, TradeSuccess)
-    assert (1 / analysis.price) == pytest.approx(Decimal("1755.115346038114345242609866"))
+    assert analysis.price == pytest.approx(Decimal(1 / 1755.115346038114345242609866))
     assert analysis.get_effective_gas_price_gwei() == 1
     assert analysis.amount_in_decimals == 6
     assert analysis.amount_out_decimals == 18